GI-103: fix: add NavDrawer component, use seperator instead of empty label, fix link
