+++ pages/medt/index.vue
@@ -25,8 +25,8 @@ const items = ref<MenuItem[]>([
 
   },
   {
-    label: '',
-    class: 'border-l border-medt-default w-1 cursor-default hover:cursor-default py-2',
+    separator: true,
+    class: 'border-l border-medt-default w-1 py-3',
   },
   {
     label: 'Netzwerktechnik',
