+++ pages/medt/multi.vue
@@ -1,12 +1,11 @@
 <script setup lang="ts">
 import type {MenuItem} from "primevue/menuitem";
+import NavDrawer from "~/components/NavDrawer.vue";
 
 useHead({
   title: "Mediendesign und Medienproduktionen | GuideIT",
   description: "Mediendesign und Medienproduktionen",
 });
-
-const visible = ref(false);
 
 const items = ref<MenuItem[]>([
   {
@@ -25,8 +24,8 @@ const items = ref<MenuItem[]>([
 
   },
   {
-    label: '',
-    class: 'border-l border-medt-mm-default w-1 cursor-default hover:cursor-default py-2',
+    separator: true,
+    class: 'border-l border-medt-mm-default w-1 py-3',
   },
   {
     label: 'Netzwerktechnik',
@@ -72,39 +71,10 @@ const items = ref<MenuItem[]>([
                    class="!rounded-3xl !bg-white-default !py-1 !px-2 !border-medt-mm-default !font-league-spartan !z-50 !shadow-medt-mm-200 !shadow-md"
                    :pt="{ item: '!focus:bg-white !active:bg-white-default !bg-none', itemContent: '!focus:bg-white-default !active:bg-white-default !bg-white-default', submenu: '!focus:bg-white-default !active:bg-white-default !bg-white-default !group-hover:underline !border-medt-default !z-50', itemLabel: '!text-black-default !text-lg group-hover:underline', itemLink: '!p-0 !py-1 !px-3'}"/>
         </div>
-        <div class="lg:hidden">
-          <Drawer v-model:visible="visible" position="right"
-                  class="!bg-white-default !text-black-default !w-72 !font-league-spartan"
-                  :pt="{header: '!pl-0 !pb-0'}">
-            <template #closeicon><i class="pi pi-angle-left !text-4xl !text-black-default"/></template>
-            <div>
-              <ul class="flex flex-col justify-end text-sub-heading-small text-end text-guideit-800">
-                <li class="mb-2">
-                  <NuxtLink to="/">Home</NuxtLink>
-                </li>
-                <li class="mb-2">
-                  <NuxtLink to="/quiz">Quiz</NuxtLink>
-                </li>
-                <li class="mb-2 text-nwt-default border-t-2 border-nwt-300 pt-2">
-                  <NuxtLink to="/nwt">Netzwerktechnik</NuxtLink>
-                </li>
-                <li class="mb-1 text-medt-default border-t-2 border-medt-300 pt-2 ">
-                  <NuxtLink to="/medt">Medientechnik</NuxtLink>
-                </li>
-                <li class="mb-1 text-text whitespace-nowrap font-light text-wrap">
-                  <NuxtLink to="/medt/multi">Mediendesign und Medienproduktionen</NuxtLink>
-                </li>
-                <li class="text-text font-light">
-                  <NuxtLink to="/medt/web">Web- und App-Entwicklung</NuxtLink>
-                </li>
-              </ul>
-            </div>
-          </Drawer>
-          <button @click="visible = true" aria-label="Open menu" aria-expanded="visible"><span
-              class="pi pi-bars !text-3xl"></span></button>
-        </div>
+        <NavDrawer/>
       </template>
-      <template #title><span class="text-medt-mm-default hyphens-auto leading-none">Mediendesign und Medienproduktionen</span></template>
+      <template #title><span
+          class="text-medt-mm-default hyphens-auto leading-none">Mediendesign und Medienproduktionen</span></template>
       <template #video>
         <div
             class="w-full aspect-video bg-medt-mm-200 border-2 rounded border-medt-mm-default flex items-center justify-center"><span
@@ -123,7 +93,7 @@ const items = ref<MenuItem[]>([
         -->
       </template>
       <template #content-1>Wenn du dich in der 3. Klasse für die Vertiefung
-        <NuxtLink to="/medt/web" class="underline text-medt-default font-medium">Medientechnik</NuxtLink>
+        <NuxtLink to="/medt" class="underline text-medt-default font-medium">Medientechnik</NuxtLink>
         ab der 4. Klasse entscheidest,
         hast du ebenfalls eine weitere wichtige Wahl zu tätigen: Deine spezielle Vertiefung in Web- und App-Entwicklung
         oder Mediendesign und Medienproduktionen. Solltest du dich für <b>Mediendesign und Medienproduktionen</b>
@@ -152,8 +122,10 @@ const items = ref<MenuItem[]>([
       </template>
       <template #content-4-header><span class="text-medt-mm-default">Mediendesign und Medienproduktionen in der Praxis erleben</span>
       </template>
-      <template #content-4>Übungen aus dem Unterricht sind zum Beispiel die Gestaltung eines <b>Brandingbooks</b>, die <b>Aufnahme
-        und Bearbeitung eines Songs</b> oder die Umsetzung einer ganzen <b>Videoproduktion</b> zu einem bestimmten Thema.
+      <template #content-4>Übungen aus dem Unterricht sind zum Beispiel die Gestaltung eines <b>Brandingbooks</b>, die
+        <b>Aufnahme
+          und Bearbeitung eines Songs</b> oder die Umsetzung einer ganzen <b>Videoproduktion</b> zu einem bestimmten
+        Thema.
 
       </template>
     </NuxtLayout>
